update routes and auth

+++ src/routes/memories.ts
@@ -1,13 +1,19 @@
 import { FastifyInstance } from "fastify";
 import { prisma } from "../lib/prisma";
 import { z } from "zod";
+
 export async function memoriesRoute(app: FastifyInstance) {
-  app.get("/users", async () => {
-    const users = await prisma.user.findMany();
-    return users;
+  app.addHook("preHandler", async (resquest) => {
+    await resquest.jwtVerify();
   });
-  app.get("/memories", async () => {
+
+
+
+  app.get("/memories", async (resquest) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: resquest.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -20,7 +26,83 @@ export async function memoriesRoute(app: FastifyInstance) {
       };
     });
   });
-  app.get("/memories/:id", async (request) => {
+
+  app.get("/memories/:id", async (request, reply) => {
+    const paramsSchema = z.object({
+      id: z.string().uuid(),
+    });
+    const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+    if (!memory.isPubluc && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    return memory;
+  });
+
+  app.post("/memories", async (request) => {
+    const bodySchema = z.object({
+      content: z.string(),
+      coverUrl: z.string(),
+      isPubluc: z.coerce.boolean().default(false),
+    });
+
+    const { content, coverUrl, isPubluc } = bodySchema.parse(request.body);
+
+    const memory = await prisma.memory.create({
+      data: {
+        content,
+        coverUrl,
+        isPubluc,
+        userId: request.user.sub,
+      },
+    });
+    return memory;
+  });
+
+  app.put("/memories/:id", async (request, reply) => {
+    const paramsSchema = z.object({
+      id: z.string().uuid(),
+    });
+    const { id } = paramsSchema.parse(request.params);
+
+    const bodySchema = z.object({
+      content: z.string(),
+      coverUrl: z.string(),
+      isPubluc: z.coerce.boolean().default(false),
+    });
+
+    const { content, coverUrl, isPubluc } = bodySchema.parse(request.body);
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
+      where: {
+        id,
+      },
+      data: {
+        content,
+        coverUrl,
+        isPubluc,
+      },
+    });
+    return memory;
+  });
+
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -32,57 +114,9 @@ export async function memoriesRoute(app: FastifyInstance) {
       },
     });
 
-    return memory;
-  });
-  app.post("/memories/", async (request) => {
-    const bodySchema = z.object({
-      content: z.string(),
-      coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
-    });
-
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
-
-    const memory = await prisma.memory.create({
-      data: {
-        content,
-        coverUrl,
-        isPublic,
-        userId: request.user.sub,
-      },
-    });
-  });
-  app.put("/memories/:id", async (request) => {
-        const paramsSchema = z.object({
-          id: z.string().uuid(),
-        });
-    const { id } = paramsSchema.parse(request.params);
-
-        const bodySchema = z.object({
-      content: z.string(),
-      coverUrl: z.string(),
-      isPublic: z.coerce.boolean().default(false),
-    });
-
-    const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
-
-    const memory = await prisma.memory.update({
-      where: {
-        id,
-      },
-      data: {
-        content,
-        coverUrl,
-        isPublic,
-      },
-    });
-return memory;
-  });
-  app.delete("/memories/:id", async () => {
-    const paramsSchema = z.object({
-      id: z.string().uuid(),
-    });
-    const { id } = paramsSchema.parse(request.params);
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
